change return type of build_key_export_request()
