+++ key-export-client/Cargo.toml
@@ -16,8 +16,10 @@ generic-ec-zkp = { git = "https://github.com/dfns-labs/generic-ec", branch = "m"
 serde_json = "1"
 rand_core = { version = "0.6", default-features = false }
 
-[target.'cfg(target_arch = "wasm32")'.dependencies]
+#[target.'cfg(target_arch = "wasm32")'.dependencies]
 wasm-bindgen = "0.2"
+serde = { version = "1.0", features = ["derive"] }
+serde-wasm-bindgen = "0.6"
 
 [dev-dependencies]
 hex = { version = "0.4", default-features = false, features = ["alloc", "serde"]}
